worker/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# InsightFace & ML
insightface==0.7.3
onnxruntime==1.17.0  # CPU version, use onnxruntime-gpu for GPU
opencv-python-headless==4.9.0.80
numpy==1.24.3
scikit-learn==1.4.0
# hdbscan==0.8.33  # Problème de compilation, on utilise DBSCAN de sklearn à la place

# Database & Storage
supabase==2.9.0
psycopg2-binary==2.9.9
httpx>=0.24.0

# Utilities
python-dotenv==1.0.0
pillow==10.2.0
requests==2.31.0
tenacity==8.2.3  # retry logic

